compose-plugin = "1.9.3 (#315)

+++ core/src/commonMain/kotlin/com/hulkdx/findprofessional/core/utils/DateUtils.kt
@@ -33,7 +33,7 @@ object DateUtils {
             .firstNotNullOf { it.key }
     }
 
-    fun lengthOfMonth(now: LocalDate): Int {
+    fun lengthOfMonth(now: LocalDate): Long {
         return now.lengthOfMonth()
     }
